+++ kroxylicious-filters/kroxylicious-authorization/src/test/java/io/kroxylicious/filter/authorization/AuthorizationFilterTest.java
@@ -300,9 +300,11 @@ class AuthorizationFilterTest {
     @Test
     void shouldSupportApis() {
         EnumSet<ApiKeys> allVersionsSupported = of(ApiKeys.API_VERSIONS,
+                ApiKeys.PRODUCE,
                 ApiKeys.SASL_HANDSHAKE,
                 ApiKeys.SASL_AUTHENTICATE,
                 ApiKeys.METADATA,
+                ApiKeys.FETCH,
                 ApiKeys.FIND_COORDINATOR,
                 ApiKeys.ADD_OFFSETS_TO_TXN,
                 ApiKeys.END_TXN,
@@ -322,7 +324,6 @@ class AuthorizationFilterTest {
                 ApiKeys.JOIN_GROUP,
                 ApiKeys.LIST_GROUPS,
                 ApiKeys.SYNC_GROUP,
-                ApiKeys.CONSUMER_GROUP_HEARTBEAT,
                 ApiKeys.DESCRIBE_CONFIGS,
                 ApiKeys.ALTER_CONFIGS,
                 ApiKeys.INCREMENTAL_ALTER_CONFIGS,
@@ -332,9 +333,8 @@ class AuthorizationFilterTest {
                 ApiKeys.DESCRIBE_TRANSACTIONS,
                 ApiKeys.HEARTBEAT,
                 ApiKeys.LIST_TRANSACTIONS);
-        EnumSet<ApiKeys> someVersionsSupported = of(ApiKeys.PRODUCE,
-                ApiKeys.FETCH,
-                ApiKeys.ADD_PARTITIONS_TO_TXN,
+        EnumSet<ApiKeys> someVersionsSupported = of(ApiKeys.ADD_PARTITIONS_TO_TXN,
+                ApiKeys.CONSUMER_GROUP_HEARTBEAT,
                 ApiKeys.INIT_PRODUCER_ID);
         EnumSet<ApiKeys> noVersionsSupported = complementOf(unionOf(allVersionsSupported, someVersionsSupported));
         for (ApiKeys apiKey : allVersionsSupported) {